Change Color to a value class with an Int

+++ core/js/src/main/scala/eu/joaocosta/minart/backend/HtmlCanvas.scala
@@ -97,9 +97,9 @@ class HtmlCanvas(val settings: Canvas.Settings) extends LowLevelCanvas {
   def getBackbufferPixel(x: Int, y: Int): Color = {
     val baseAddr = 4 * (y * settings.scale * settings.scaledWidth + (x * settings.scale))
     Color(
-      buffer.data(baseAddr + 0).toShort,
-      buffer.data(baseAddr + 1).toShort,
-      buffer.data(baseAddr + 2).toShort)
+      buffer.data(baseAddr + 0),
+      buffer.data(baseAddr + 1),
+      buffer.data(baseAddr + 2))
   }
 
   def getBackbuffer(): Vector[Vector[Color]] = {
@@ -108,7 +108,7 @@ class HtmlCanvas(val settings: Canvas.Settings) extends LowLevelCanvas {
       val lineBase = y * settings.scale * settings.scaledWidth
       columns.map { x =>
         val baseAddr = 4 * (lineBase + (x * settings.scale))
-        Color(imgData(baseAddr).toShort, imgData(baseAddr + 1).toShort, imgData(baseAddr + 2).toShort)
+        Color(imgData(baseAddr), imgData(baseAddr + 1), imgData(baseAddr + 2))
       }.toVector
     }.toVector
   }
